chat_service/requirements.txt
fastapi
uvicorn[standard]
httpx
websockets
google.generativeai